Splitting cache_test into two for Windows
